Move `polygon/bor/valset` to `polygon/heimdall` and flatten (#16556)
