+++ polygon/bor/bor_internal_test.go
@@ -30,7 +30,6 @@ import (
 	"github.com/erigontech/erigon-lib/types"
 	"github.com/erigontech/erigon/execution/consensus"
 	"github.com/erigontech/erigon/polygon/bor/statefull"
-	"github.com/erigontech/erigon/polygon/bor/valset"
 	polychain "github.com/erigontech/erigon/polygon/chain"
 	"github.com/erigontech/erigon/polygon/heimdall"
 )
@@ -59,7 +58,7 @@ func (m mockSpanReader) Span(context.Context, uint64) (*heimdall.Span, bool, err
 	panic("mock")
 }
 
-func (m mockSpanReader) Producers(context.Context, uint64) (*valset.ValidatorSet, error) {
+func (m mockSpanReader) Producers(context.Context, uint64) (*heimdall.ValidatorSet, error) {
 	panic("mock")
 }
