+++ polygon/heimdall/client_idle.go
@@ -22,7 +22,6 @@ import (
 	"time"
 
 	"github.com/erigontech/erigon/params"
-	"github.com/erigontech/erigon/polygon/bor/valset"
 )
 
 type IdleClient struct {
@@ -35,8 +34,8 @@ func NewIdleClient(cfg params.MiningConfig) Client {
 
 func (c *IdleClient) FetchLatestSpan(ctx context.Context) (*Span, error) {
 	return &Span{
-		ValidatorSet: valset.ValidatorSet{
-			Validators: []*valset.Validator{
+		ValidatorSet: ValidatorSet{
+			Validators: []*Validator{
 				{
 					ID:          0,
 					Address:     c.cfg.Etherbase,
@@ -44,7 +43,7 @@ func (c *IdleClient) FetchLatestSpan(ctx context.Context) (*Span, error) {
 				},
 			},
 		},
-		SelectedProducers: []valset.Validator{
+		SelectedProducers: []Validator{
 			{
 				ID:          0,
 				Address:     c.cfg.Etherbase,
@@ -57,8 +56,8 @@ func (c *IdleClient) FetchLatestSpan(ctx context.Context) (*Span, error) {
 func (c *IdleClient) FetchSpan(ctx context.Context, spanID uint64) (*Span, error) {
 	return &Span{
 		Id: SpanId(spanID),
-		ValidatorSet: valset.ValidatorSet{
-			Validators: []*valset.Validator{
+		ValidatorSet: ValidatorSet{
+			Validators: []*Validator{
 				{
 					ID:          0,
 					Address:     c.cfg.Etherbase,
@@ -66,7 +65,7 @@ func (c *IdleClient) FetchSpan(ctx context.Context, spanID uint64) (*Span, error
 				},
 			},
 		},
-		SelectedProducers: []valset.Validator{
+		SelectedProducers: []Validator{
 			{
 				ID:          0,
 				Address:     c.cfg.Etherbase,
