+++ polygon/heimdall/service.go
@@ -29,7 +29,6 @@ import (
 	"github.com/erigontech/erigon-lib/event"
 	"github.com/erigontech/erigon-lib/log/v3"
 	"github.com/erigontech/erigon/polygon/bor/borcfg"
-	"github.com/erigontech/erigon/polygon/bor/valset"
 	"github.com/erigontech/erigon/polygon/heimdall/poshttp"
 )
 
@@ -229,7 +228,7 @@ func (s *Service) MilestonesFromBlock(ctx context.Context, startBlock uint64) ([
 	return s.reader.MilestonesFromBlock(ctx, startBlock)
 }
 
-func (s *Service) Producers(ctx context.Context, blockNum uint64) (*valset.ValidatorSet, error) {
+func (s *Service) Producers(ctx context.Context, blockNum uint64) (*ValidatorSet, error) {
 	return s.reader.Producers(ctx, blockNum)
 }
