+++ polygon/heimdall/span_block_producers_tracker.go
@@ -27,7 +27,6 @@ import (
 
 	"github.com/erigontech/erigon-lib/log/v3"
 	"github.com/erigontech/erigon/polygon/bor/borcfg"
-	"github.com/erigontech/erigon/polygon/bor/valset"
 )
 
 func newSpanBlockProducersTracker(
@@ -124,7 +123,7 @@ func (t *spanBlockProducersTracker) ObserveSpan(ctx context.Context, newSpan *Sp
 			EndBlock:   newSpan.EndBlock,
 			// https://github.com/maticnetwork/genesis-contracts/blob/master/contracts/BorValidatorSet.template#L82-L89
 			// initial producers == initial validators
-			Producers: valset.NewValidatorSet(newSpan.ValidatorSet.Validators),
+			Producers: NewValidatorSet(newSpan.ValidatorSet.Validators),
 		}
 		err = t.store.PutEntity(ctx, uint64(newProducerSelection.SpanId), newProducerSelection)
 		if err != nil {
@@ -158,11 +157,11 @@ func (t *spanBlockProducersTracker) ObserveSpan(ctx context.Context, newSpan *Sp
 	spanEndSprintNum := t.borConfig.CalculateSprintNumber(lastProducerSelection.EndBlock)
 	increments := int(spanEndSprintNum - spanStartSprintNum)
 	for i := 0; i < increments; i++ {
-		producers = valset.GetUpdatedValidatorSet(producers, producers.Validators, t.logger)
+		producers = GetUpdatedValidatorSet(producers, producers.Validators, t.logger)
 		producers.IncrementProposerPriority(1)
 	}
 
-	newProducers := valset.GetUpdatedValidatorSet(producers, newSpan.Producers(), t.logger)
+	newProducers := GetUpdatedValidatorSet(producers, newSpan.Producers(), t.logger)
 	newProducers.IncrementProposerPriority(1)
 	newProducerSelection := &SpanBlockProducerSelection{
 		SpanId:     newSpan.Id,
@@ -179,7 +178,7 @@ func (t *spanBlockProducersTracker) ObserveSpan(ctx context.Context, newSpan *Sp
 	return nil
 }
 
-func (t *spanBlockProducersTracker) Producers(ctx context.Context, blockNum uint64) (*valset.ValidatorSet, error) {
+func (t *spanBlockProducersTracker) Producers(ctx context.Context, blockNum uint64) (*ValidatorSet, error) {
 	startTime := time.Now()
 
 	producers, increments, err := t.producers(ctx, blockNum)
@@ -197,7 +196,7 @@ func (t *spanBlockProducersTracker) Producers(ctx context.Context, blockNum uint
 	return producers, nil
 }
 
-func (t *spanBlockProducersTracker) producers(ctx context.Context, blockNum uint64) (*valset.ValidatorSet, int, error) {
+func (t *spanBlockProducersTracker) producers(ctx context.Context, blockNum uint64) (*ValidatorSet, int, error) {
 	currentSprintNum := t.borConfig.CalculateSprintNumber(blockNum)
 
 	// have we previously calculated the producers for the same sprint num (chain tip optimisation)
@@ -239,7 +238,7 @@ func (t *spanBlockProducersTracker) producers(ctx context.Context, blockNum uint
 	spanStartSprintNum := t.borConfig.CalculateSprintNumber(producerSelection.StartBlock)
 	increments := int(currentSprintNum - spanStartSprintNum)
 	for i := 0; i < increments; i++ {
-		producers = valset.GetUpdatedValidatorSet(producers, producers.Validators, t.logger)
+		producers = GetUpdatedValidatorSet(producers, producers.Validators, t.logger)
 		producers.IncrementProposerPriority(1)
 	}
