+++ polygon/heimdall/validator_set.go
@@ -14,9 +14,7 @@
 // You should have received a copy of the GNU Lesser General Public License
 // along with Erigon. If not, see <http://www.gnu.org/licenses/>.
 
-package valset
-
-// Tendermint leader selection algorithm
+package heimdall
 
 import (
 	"bytes"
@@ -45,6 +43,161 @@ const (
 	MaxTotalVotingPower      = int64(math.MaxInt64) / 8
 	PriorityWindowSizeFactor = 2
 )
+
+type Validator struct {
+	ID               uint64         `json:"ID"`
+	Address          common.Address `json:"signer"`
+	VotingPower      int64          `json:"power"`
+	ProposerPriority int64          `json:"accum"`
+}
+
+// NewValidator creates new validator
+func NewValidator(address common.Address, votingPower int64) *Validator {
+	return &Validator{
+		Address:          address,
+		VotingPower:      votingPower,
+		ProposerPriority: 0,
+	}
+}
+
+// Copy creates a new copy of the validator so we can mutate ProposerPriority.
+// Panics if the validator is nil.
+func (v *Validator) Copy() *Validator {
+	vCopy := *v
+	return &vCopy
+}
+
+// Cmp returns the one validator with a higher ProposerPriority.
+// If ProposerPriority is same, it returns the validator with lexicographically smaller address
+func (v *Validator) Cmp(other *Validator) *Validator {
+	// if both of v and other are nil, nil will be returned and that could possibly lead to nil pointer dereference bubbling up the stack
+	if v == nil {
+		return other
+	}
+
+	if other == nil {
+		return v
+	}
+
+	if v.ProposerPriority > other.ProposerPriority {
+		return v
+	}
+
+	if v.ProposerPriority < other.ProposerPriority {
+		return other
+	}
+
+	result := bytes.Compare(v.Address.Bytes(), other.Address.Bytes())
+
+	if result == 0 {
+		panic("Cannot compare identical validators")
+	}
+
+	if result < 0 {
+		return v
+	}
+
+	// result > 0
+	return other
+}
+
+func (v *Validator) String() string {
+	if v == nil {
+		return "nil-Validator"
+	}
+
+	return fmt.Sprintf("Validator{%v ID: %v Power:%v Priority:%v}",
+		v.Address.Hex(),
+		v.ID,
+		v.VotingPower,
+		v.ProposerPriority)
+}
+
+// ValidatorListString returns a prettified validator list for logging purposes.
+func ValidatorListString(vals []*Validator) string {
+	chunks := make([]string, len(vals))
+	for i, val := range vals {
+		chunks[i] = fmt.Sprintf("%s:%d", val.Address, val.VotingPower)
+	}
+
+	return strings.Join(chunks, ",")
+}
+
+// HeaderBytes return header bytes
+func (v *Validator) HeaderBytes() []byte {
+	result := make([]byte, 40)
+	copy(result[:20], v.Address.Bytes())
+	copy(result[20:], v.PowerBytes())
+
+	return result
+}
+
+// PowerBytes return power bytes
+func (v *Validator) PowerBytes() []byte {
+	powerBytes := big.NewInt(0).SetInt64(v.VotingPower).Bytes()
+	result := make([]byte, 20)
+	copy(result[20-len(powerBytes):], powerBytes)
+
+	return result
+}
+
+// MinimalVal returns block number of last validator update
+func (v *Validator) MinimalVal() MinimalVal {
+	return MinimalVal{
+		ID:          v.ID,
+		VotingPower: uint64(v.VotingPower),
+		Signer:      v.Address,
+	}
+}
+
+// ParseValidators returns validator set bytes
+func ParseValidators(validatorsBytes []byte) ([]*Validator, error) {
+	if len(validatorsBytes)%40 != 0 {
+		return nil, errors.New("invalid validators bytes")
+	}
+
+	result := make([]*Validator, len(validatorsBytes)/40)
+
+	for i := 0; i < len(validatorsBytes); i += 40 {
+		address := make([]byte, 20)
+		power := make([]byte, 20)
+
+		copy(address, validatorsBytes[i:i+20])
+		copy(power, validatorsBytes[i+20:i+40])
+
+		result[i/40] = NewValidator(common.BytesToAddress(address), big.NewInt(0).SetBytes(power).Int64())
+	}
+
+	return result, nil
+}
+
+// ---
+
+// MinimalVal is the minimal validator representation
+// Used to send validator information to bor validator contract
+type MinimalVal struct {
+	ID          uint64         `json:"ID"`
+	VotingPower uint64         `json:"power"` // TODO add 10^-18 here so that we don't overflow easily
+	Signer      common.Address `json:"signer"`
+}
+
+// SortMinimalValByAddress sorts validators
+func SortMinimalValByAddress(a []MinimalVal) []MinimalVal {
+	sort.Slice(a, func(i, j int) bool {
+		return bytes.Compare(a[i].Signer.Bytes(), a[j].Signer.Bytes()) < 0
+	})
+
+	return a
+}
+
+// ValidatorsToMinimalValidators converts array of validators to minimal validators
+func ValidatorsToMinimalValidators(vals []Validator) (minVals []MinimalVal) {
+	for _, val := range vals {
+		minVals = append(minVals, val.MinimalVal())
+	}
+
+	return
+}
 
 // ValidatorSet represent a set of *Validator at a given height.
 // The validators can be fetched by address or index.
@@ -880,4 +1033,61 @@ func validatorContains(a []*Validator, x *Validator) (*Validator, bool) {
 	}
 
 	return nil, false
+}
+
+// TotalVotingPowerExceededError is returned when the maximum allowed total voting power is exceeded
+type TotalVotingPowerExceededError struct {
+	Sum        int64
+	Validators []*Validator
+}
+
+func (e *TotalVotingPowerExceededError) Error() string {
+	return fmt.Sprintf(
+		"Total voting power should be guarded to not exceed %v; got: %v; for validator set: %v",
+		MaxTotalVotingPower,
+		e.Sum,
+		e.Validators,
+	)
+}
+
+type InvalidStartEndBlockError struct {
+	Start         uint64
+	End           uint64
+	CurrentHeader uint64
+}
+
+func (e *InvalidStartEndBlockError) Error() string {
+	return fmt.Sprintf(
+		"Invalid parameters start: %d and end block: %d params",
+		e.Start,
+		e.End,
+	)
+}
+
+// UnauthorizedProposerError is returned if a header is [being] signed by an unauthorized entity.
+type UnauthorizedProposerError struct {
+	Number   uint64
+	Proposer []byte
+}
+
+func (e *UnauthorizedProposerError) Error() string {
+	return fmt.Sprintf(
+		"Proposer 0x%x is not a part of the producer set at block %d",
+		e.Proposer,
+		e.Number,
+	)
+}
+
+// UnauthorizedSignerError is returned if a header is [being] signed by an unauthorized entity.
+type UnauthorizedSignerError struct {
+	Number uint64
+	Signer []byte
+}
+
+func (e *UnauthorizedSignerError) Error() string {
+	return fmt.Sprintf(
+		"Signer 0x%x is not a part of the producer set at block %d",
+		e.Signer,
+		e.Number,
+	)
 }
